[US0009]: Light code cleanup
