+++ scienceorfiction.env
@@ -1,3 +1,5 @@
+# public environment variables
+
 # flask settings
 FLASK_APP=wsgi.py
 FLASK_ENV=docker
