+++ scienceorfiction/wsgi.py
@@ -1,3 +1,5 @@
+# Standard App Factory setup
+
 from app import create_app
 
 app = create_app()
